[129/webpack] add babel/webpack to avoid es5-ing
