+++ BookReaderDemo/demo-ia-plato.html
@@ -28,7 +28,6 @@
     <script src="../BookReader/plugins/plugin.url.js"></script>
     <script src="../BookReader/plugins/plugin.resume.js"></script>
     <script src="../BookReader/plugins/plugin.chapters.js"></script>
-    <script src="../BookReader/utils/AsyncStream.js"></script>
     <script src="../BookReader/plugins/plugin.tts.js"></script>
     <script src="../BookReader/plugins/plugin.search.js"></script>
