+++ webpack.config.js
@@ -1,0 +1,15 @@
+const path = require('path');
+
+module.exports = {
+    mode: 'production',
+    entry: './src/js/plugins/plugin.tts.js',
+    module: {
+        rules: [
+            { test: /\.js$/, exclude: /node_modules/, loader: "babel-loader" }
+        ]
+    },
+    output: {
+        filename: 'plugin.tts.js',
+        path: path.resolve(__dirname, 'BookReader/plugins')
+    }
+};
